reorganized controller code
